added depth to the script
